+++ src/main/kotlin/shop/itbug/fluttercheckversionx/util/MyFileUtil.kt
@@ -1,15 +1,24 @@
 package shop.itbug.fluttercheckversionx.util
 
 import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer
+import com.intellij.json.JsonFileType
 import com.intellij.openapi.application.runReadAction
+import com.intellij.openapi.command.WriteCommandAction
+import com.intellij.openapi.fileEditor.FileEditorManager
+import com.intellij.openapi.progress.ProgressIndicator
+import com.intellij.openapi.progress.Task
+import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.guessProjectDir
+import com.intellij.openapi.startup.StartupManager
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.psi.PsiManager
+import com.intellij.psi.codeStyle.CodeStyleManager
 import com.intellij.psi.search.FileTypeIndex
 import com.intellij.psi.search.GlobalSearchScopes
+import com.intellij.testFramework.LightVirtualFile
 import com.intellij.util.indexing.FileBasedIndex
 import com.jetbrains.lang.dart.DartFileType
 import org.jetbrains.yaml.psi.YAMLFile
@@ -101,7 +110,39 @@ object MyFileUtil {
      */
     fun reIndexPubspecFile(project: Project) {
         getPubspecVirtualFile(project)?.let { virtualFile ->
-            FileBasedIndex.getInstance().requestReindex(virtualFile)
+            if (checkFileIsIndex(project, virtualFile)) {
+                println("reindex pubspec.yaml")
+                FileBasedIndex.getInstance().requestReindex(virtualFile)
+            }
+        }
+    }
+
+
+    /**
+     * 检查文件是否被索引?
+     */
+    fun checkFileIsIndex(project: Project, file: VirtualFile): Boolean {
+        try {
+            val file = runReadAction { PsiManager.getInstance(project).findFile(file) }
+            return file != null
+        } catch (_: Exception) {
+            return false
+        }
+        return false
+    }
+
+    /**
+     * 重新索引指定文件
+     */
+    fun reIndexFile(project: Project, file: VirtualFile) {
+        StartupManager.getInstance(project).runAfterOpened {
+            DumbService.getInstance(project).runWhenSmart {
+                if (checkFileIsIndex(project, file)) {
+                    println("reindex = ${file.name}")
+                    FileBasedIndex.getInstance().requestReindex(file)
+                }
+
+            }
         }
     }
 
@@ -116,4 +157,43 @@ object MyFileUtil {
         val files = FileTypeIndex.getFiles(DartFileType.INSTANCE, scope)
         return files.toList()
     }
+
+    /**
+     * 创建一个json的虚拟文件
+     */
+    fun createVirtualFileByJsonText(
+        text: String,
+        filename: String,
+        action: (file: VirtualFile, tool: MyFileUtil) -> Unit
+    ): VirtualFile {
+        val vf = LightVirtualFile(filename, JsonFileType.INSTANCE, text)
+        action.invoke(vf, this)
+        return vf
+    }
+
+    /**
+     * 打开某个文件
+     */
+    fun openInEditor(file: VirtualFile, project: Project) {
+        FileEditorManager.getInstance(project).openFile(file, true)
+    }
+
+    /**
+     * 格式化文件
+     */
+    fun reformatVirtualFile(file: VirtualFile, project: Project) {
+        val vf = runReadAction { PsiManager.getInstance(project).findFile(file) }
+        if (vf != null) {
+            val task = object : Task.Backgroundable(project, "Reformat ${file.name}", false) {
+                override fun run(p0: ProgressIndicator) {
+                    WriteCommandAction.runWriteCommandAction(project) {
+                        CodeStyleManager.getInstance(project).reformat(vf)
+                    }
+                }
+            }
+            task.queue()
+
+        }
+    }
+
 }